+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:theme="@style/Theme.Ddaraogae"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.ui.home.AllStampActivity"
+            android:name=".presentation.ui.stamp.AllStampActivity"
             android:exported="false" />
 
         <meta-data
@@ -32,19 +32,19 @@
             android:value="@string/ddaraogae_client_id" />
 
         <activity
-            android:name=".presentation.ui.home.AddActivity"
+            android:name=".presentation.ui.add.AddActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.ui.home.HistoryActivity"
+            android:name=".presentation.ui.history.HistoryActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.ui.walk.FinishActivity"
+            android:name=".presentation.ui.finish.FinishActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.ui.mypage.EditPetActivity"
+            android:name=".presentation.ui.edit.EditPetActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.ui.mypage.DetailPetActivity"
+            android:name=".presentation.ui.dog.DetailPetActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.ui.main.MainActivity"
