+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/add/AddActivity.kt
@@ -1,8 +1,7 @@
-package com.nbcfinalteam2.ddaraogae.presentation.ui.home
+package com.nbcfinalteam2.ddaraogae.presentation.ui.add
 
 import android.Manifest
 import android.content.Intent
-import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.view.View
