+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/dog/DetailUiState.kt
@@ -1,4 +1,4 @@
-package com.nbcfinalteam2.ddaraogae.presentation.ui.mypage
+package com.nbcfinalteam2.ddaraogae.presentation.ui.dog
 
 data class DetailUiState(
     val isLoading: Boolean
