+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/finish/FinishActivity.kt
@@ -1,4 +1,4 @@
-package com.nbcfinalteam2.ddaraogae.presentation.ui.walk
+package com.nbcfinalteam2.ddaraogae.presentation.ui.finish
 
 import android.Manifest
 import android.content.pm.PackageManager
@@ -30,7 +30,7 @@ import com.nbcfinalteam2.ddaraogae.presentation.model.DefaultEvent
 import com.nbcfinalteam2.ddaraogae.presentation.model.DogInfo
 import com.nbcfinalteam2.ddaraogae.presentation.model.WalkingInfo
 import com.nbcfinalteam2.ddaraogae.presentation.ui.loading.LoadingDialog
-import com.nbcfinalteam2.ddaraogae.presentation.ui.walk.StampDialogFragment.Companion.ARG_STAMP_LIST
+import com.nbcfinalteam2.ddaraogae.presentation.ui.finish.StampDialogFragment.Companion.ARG_STAMP_LIST
 import com.nbcfinalteam2.ddaraogae.presentation.util.ImageConverter.bitmapToByteArray
 import com.nbcfinalteam2.ddaraogae.presentation.util.TextConverter.dateDateToString
 import com.nbcfinalteam2.ddaraogae.presentation.util.TextConverter.distanceDoubleToString
