+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/finish/FinishUiState.kt
@@ -1,4 +1,4 @@
-package com.nbcfinalteam2.ddaraogae.presentation.ui.walk
+package com.nbcfinalteam2.ddaraogae.presentation.ui.finish
 
 data class FinishUiState(
     val isLoading: Boolean
