+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/finish/StampDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.nbcfinalteam2.ddaraogae.presentation.ui.walk
+package com.nbcfinalteam2.ddaraogae.presentation.ui.finish
 
 import android.os.Bundle
 import android.view.LayoutInflater
