+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/history/HistoryOnClickListener.kt
@@ -1,4 +1,4 @@
-package com.nbcfinalteam2.ddaraogae.presentation.ui.home
+package com.nbcfinalteam2.ddaraogae.presentation.ui.history
 
 interface HistoryOnClickListener {
     fun onMonthClick(year: Int, monthNumber: Int)
