+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/walk/WalkFragment.kt
@@ -37,9 +37,10 @@ import com.nbcfinalteam2.ddaraogae.presentation.model.DogInfo
 import com.nbcfinalteam2.ddaraogae.presentation.model.WalkingInfo
 import com.nbcfinalteam2.ddaraogae.presentation.service.LocationService
 import com.nbcfinalteam2.ddaraogae.presentation.service.ServiceInfoState
-import com.nbcfinalteam2.ddaraogae.presentation.ui.walk.FinishActivity.Companion.LOCATIONLIST
-import com.nbcfinalteam2.ddaraogae.presentation.ui.walk.FinishActivity.Companion.WALKINGDOGS
-import com.nbcfinalteam2.ddaraogae.presentation.ui.walk.FinishActivity.Companion.WALKINGUIMODEL
+import com.nbcfinalteam2.ddaraogae.presentation.ui.finish.FinishActivity
+import com.nbcfinalteam2.ddaraogae.presentation.ui.finish.FinishActivity.Companion.LOCATIONLIST
+import com.nbcfinalteam2.ddaraogae.presentation.ui.finish.FinishActivity.Companion.WALKINGDOGS
+import com.nbcfinalteam2.ddaraogae.presentation.ui.finish.FinishActivity.Companion.WALKINGUIMODEL
 import com.nbcfinalteam2.ddaraogae.presentation.util.TextConverter.distanceDoubleToString
 import com.nbcfinalteam2.ddaraogae.presentation.util.TextConverter.timeIntToString
 import com.nbcfinalteam2.ddaraogae.presentation.util.ToastMaker
