+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/util/GraphUtils.kt
@@ -1,4 +1,4 @@
-package com.nbcfinalteam2.ddaraogae.presentation.ui.home
+package com.nbcfinalteam2.ddaraogae.presentation.util
 
 import android.content.Context
 import com.github.mikephil.charting.charts.LineChart
@@ -7,7 +7,6 @@ import com.github.mikephil.charting.components.YAxis
 import com.github.mikephil.charting.data.LineData
 import com.github.mikephil.charting.formatter.ValueFormatter
 import com.nbcfinalteam2.ddaraogae.R
-import com.nbcfinalteam2.ddaraogae.presentation.util.DateFormatter
 
 object GraphUtils {
