+++ app/src/main/res/layout/activity_add.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.ui.home.AddActivity">
+    tools:context=".presentation.ui.add.AddActivity">
 
     <ScrollView
         android:layout_width="match_parent"
