+++ app/src/main/res/layout/activity_detail_pet.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.ui.mypage.DetailPetActivity">
+    tools:context=".presentation.ui.dog.DetailPetActivity">
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_appbar"
         android:layout_width="wrap_content"
