+++ app/src/main/res/layout/activity_edit_pet.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".presentation.ui.mypage.EditPetActivity">
+    tools:context=".presentation.ui.edit.EditPetActivity">
 
     <ScrollView
         android:layout_width="match_parent"
