+++ app/src/main/res/layout/activity_finish.xml
@@ -6,7 +6,7 @@
     android:background="@color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.walk.FinishActivity">
+    tools:context=".presentation.ui.finish.FinishActivity">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/tv_finish_title"
